adding other standardized level functions and splitting them into files. Lacking docs and tests yet
