+++ pyfar/level/__init__.py
@@ -2,10 +2,22 @@
 
 from .time_weighting import (
     time_weighted_sound_pressure,
+)
+from .standard_levels import (
     time_weighted_level,
+    equivalent_continuous_level,
+    sliding_equivalent_continuous_level,
+    exposure_level,
+    peak_level,
+    maximum_time_weighted_level,
 )
 
 __all__ = [
     'time_weighted_sound_pressure',
     'time_weighted_level',
+    'equivalent_continuous_level',
+    'sliding_equivalent_continuous_level',
+    'exposure_level',
+    'peak_level',
+    'maximum_time_weighted_level',
 ]
